+++ src/lib.rs
@@ -1,5 +1,5 @@
 mod auth;
-mod client_v2;
+mod client;
 mod connection;
 mod consumer;
 mod consumer_engine;
@@ -10,6 +10,6 @@ mod producer;
 mod resolver_manager;
 
 pub use auth::*;
-pub use client_v2::*;
+pub use client::*;
 pub use consumer::*;
 pub use producer::*;
